% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{add_standards}
\alias{add_standards}
\title{Add calibration standards}
\usage{
add_standards(dt, stds, match_by = default(match_by),
  is_standard = default(is_standard), quiet = setting(quiet))
}
\arguments{
\item{dt}{data table}

\item{stds}{the standards table}
}
\description{
Add calibration standards
}
