% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table.R
\name{aggregate_peak_table}
\alias{aggregate_peak_table}
\title{Aggregate peak table
NOTE: don't use this - operate independently of isoreader structure!}
\usage{
aggregate_peak_table(isofiles, with_units = TRUE, select = all_columns(),
  include_file_info = c(), quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{with_units}{whether to include units in the column headers (if there are any) or not}

\item{select}{which vendor table columns select. All by default.}
}
\description{
Aggregate data from the peak table. This information is only available once a peak table is generated (see e.g. \code{\link{use_vendor_identified_peaks}}).
}
