% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calculations.R
\name{calculate_area}
\alias{calculate_area}
\title{calculates area of peak from discrete signal and time measurements
can deal with time points out of order as long as time and signal indices are matched}
\usage{
calculate_area(t, y, tstart = NULL, tend = NULL, const_dt = FALSE)
}
\arguments{
\item{t}{the time points}

\item{y}{the signal data}

\item{tstart}{the starting time of the signal to integrate, NULL (default) starts with the signal from the first time point}

\item{tend}{the ending time of the signal to integrate, NULL (default) ends with the signal at the last time point}

\item{const_dt}{if the time interval between data points is knonw to be constant use TRUE to make the computation slightly faster (~2-3 times with peaks 10,000 data points wide), default is FALSE}
}
\value{
integrated area in units of [y units][t units]
}
\description{
calculates area of peak from discrete signal and time measurements
can deal with time points out of order as long as time and signal indices are matched
}
