% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calculations.R
\name{calculate_backgrounds}
\alias{calculate_backgrounds}
\title{Calculate background signals}
\usage{
calculate_backgrounds(data, method, signal_pattern = "^[vi]\\\\d+",
  bg_suffix = "_bgrd", time = default("time"),
  peak = default("peak"), bgrd = default("bgrd"),
  file = default("file"), quiet = default("quiet"))
}
\arguments{
\item{data}{the data frame with chromatographic and peak information}

\item{method}{what method to use to calculate backgrounds (backgrounds are ALWAYS based solely on time points marked as background)
run_const = constant background averaged across entire run
run_linear = linear background across run @FIXME NOT IMPELEMENTED YET
peak_const = constant background across single peak @FIXME NOT IMPELEMENTED YET
peak_linear = linear background across peak @FIXME NOT IMPELEMENTED YET}

\item{signal_pattern}{pattern for signal columns (by default detects voltages starting with e.g. v45 and currents starting with e.g. i45)}

\item{bg_suffix}{the suffix for the new background columns}

\item{time}{column name of the time column to use to find peak centers}

\item{peak}{column name of the peak id column that identifies different peaks}

\item{file}{column name of the filename column}
}
\value{
data frame with new signal columns with bg_suffix added, as well as a p.calc_bgrd_method
}
\description{
This function calculates peak backgrounds from a chromatographic trace with peak table data (see \code{\link{iso_combine_raw_data_with_peak_table}}).
}
\note{
perhaps implement possibilty to pass with an actual function as method that takes the sub data frame as parameter?
}
