% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{evaluate_range}
\alias{evaluate_range}
\title{evaluate regression ranges}
\usage{
evaluate_range(dt, ..., nested_model = FALSE, model_data = model_data,
  model_params = model_params, in_reg = in_reg,
  model_range = model_range, in_range = in_range)
}
\arguments{
\item{dt}{data table}

\item{...}{which terms to evaluate the range for. Can be individual columns or more complex numeric expressions. All must be valid within the scope of the model_data and only numeric values are currently evaluated.}

\item{nested_model}{whether the model is nested, if TRUE, must also provide \code{model_params}}

\item{model_data}{the nested model data column}

\item{model_params}{the nested model information (only relevant if \code{nest_model = TRUE})}

\item{in_reg}{name of the new logical column in the nested model_data that flags the data included in the regression (i.e. the data that fullfills the \code{model_filter_condition} if provided)}

\item{model_range}{new column recording the range of all terms for each model and model_data combination.}

\item{in_range}{new column in the model_data that holds whether a data entry is within the model_range. Checks whether all terms are within the model_range and records a textual summary of the result in this new column.}
}
\description{
evaluate regression ranges
}
