% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_standards_calibrations}
\alias{get_standards_calibrations}
\title{Calculate calibrations.}
\usage{
get_standards_calibrations(dt, stds, match_by = default(match_by),
  group_by = default(group_by), measured = default(measured),
  known = default(known), ampl = default(ampl), keep = default(NULL))
}
\arguments{
\item{dt}{the data table}

\item{stds}{the standards table}

\item{match_by}{what to use to match the standards to the data (use c(...) to select multiple)}

\item{group_by}{what to group by for indidual calibration calculations (use c(...) to select multiple)}

\item{measured}{where is the measured data stored?}

\item{known}{where is the known data value stored?}

\item{ampl}{where is the signal amplitude stored?}

\item{keep}{which column(s) to keep beyond the group_by (use c(...) to select multiple).
Note that if these should be unique across grouping variables, otherwise it will result in multiple lines of the same regression data per group.}
}
\description{
Calculate calibrations.
}
\note{
: todo - find a better way to deal with the amplitude part

: todo - deal with the keep columns better so they don't end up working as grouping columns during nesting! maybe join back in after the operations?
}
