% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{iso_add_metadata}
\alias{iso_add_metadata}
\title{Add metadata to data table}
\usage{
iso_add_metadata(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function was deprecated in favour of the more broadly applicable \code{\link[isoreader]{iso_add_file_info}}.
}
