% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_add_standards}
\alias{iso_add_standards}
\title{Add calibration standards}
\usage{
iso_add_standards(dt, stds, match_by = default(std_match_by),
  is_std_peak = default(is_std_peak), quiet = default(quiet),
  is_standard = NULL)
}
\arguments{
\item{dt}{data table}

\item{stds}{standards data frame}

\item{match_by}{what column(s) to match the standards by}

\item{is_std_peak}{new column that holds information about which ones are standard peaks (i.e. have known isotopic values)}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}

\item{is_standard}{renamed to \code{is_std_peak} because the naming caused too much confusion, will be removed in future versions, please use \code{is_std_peak} instead}
}
\value{
data frame with standards data frame merged in and the following information column added:
\itemize{
\item{\code{is standard}: }{a logical TRUE/FALSE indicating which data table entry is a standard}
}
}
\description{
Add calibration standards
}
\concept{calibration functions}
