% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_calibrate_delta}
\alias{iso_calibrate_delta}
\title{Calibrate delta values}
\usage{
iso_calibrate_delta(dt, model, is_standard = default(is_standard),
  calibration_data = default(calibration_data),
  delta_residual = default(delta_residual), quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{model}{one or more models to run delta calibration with}

\item{is_standard}{column or filter condition to determine which data to use for the calibration (default is the field introduced by \code{\link{iso_prepare_for_calibration}})}

\item{calibration_data}{column name for the combined calibration data}

\item{delta_residual}{name of the new column for keeping track of the residuals of the calibration model}
}
\description{
Before running this function, make sure to group the data appropriately by running \code{\link{iso_prepare_for_calibration}}.
}
