% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_evaluate_calibration_range}
\alias{iso_evaluate_calibration_range}
\title{Evaluate calibration range}
\usage{
iso_evaluate_calibration_range(dt, ..., calibration = "",
  quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{...}{which terms to evaluate the range for. Can be individual columns or more complex numeric expressions. All must be valid within the scope of the model_data and only numeric values are currently evaluated.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\description{
Evaluates the calibration ranges for all calibrations and all data with respect to the provided terms (\code{...}). Generates a summary column called \code{in_range} (with \code{calibration} prefix if used) in the \code{all_data} data frames summarizing the range information. Also stores the calibration ranges themselves in a nested data frame, which can be accessed via \link{iso_unnest_calibration_range} if needed.
}
\details{
Note that this function requires prior generation of a calibration (\code{\link{iso_generate_calibration}}). All measured parameters and derived terms can be included in the calibration range evalution. However, if the predicted term is intended to be included in the range evaluation, the calibration(s) must also be applied (\code{\link{iso_apply_calibration}}) first so the predicted term is actually available.
}
