% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_calibration_to_excel}
\alias{iso_export_calibration_to_excel}
\title{Export calibration to Excel}
\usage{
iso_export_calibration_to_excel(calibs_df, filepath,
  include_all_data = TRUE, include_model_coefs = TRUE,
  include_model_summary = TRUE, include_model_range = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{calibs_df}{calibration(s) data frame (i.e. the output of functions like \link{iso_generate_calibration}, \link{iso_apply_calibration}, \link{iso_evaluate_calibration_range}). Must still be nested (i.e. before running any of the \code{iso_unnest...} functions) to make sure it contains alll information.}

\item{filepath}{the path where to store the Excel file.}

\item{include_all_data}{whether to include a tab for all the peak data - if this function is called after \link{iso_apply_calibration} this will include}

\item{include_model_coefs}{whether to include a tab for the model coefficients}

\item{include_model_summary}{whether to include a tab for the model summary}

\item{include_model_range}{whether to include a tab for the evaluation range (if available, i.e. generated by \link{iso_evaluate_calibration_range})}
}
\value{
returns the \code{calib} object invisibly for use in pipelines
}
\description{
This function exports the passed in calibration(s) data frame to Excel. The different kinds of information (all data, model coefficients, summary, evaluation range) are exported to separate tabs within the excel file. If there are multiple different calibrations in the data frame (i.e. multiple \code{*calib_params} columns), ALL will be exported with different tabs for each.
}
