% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_data_to_excel}
\alias{iso_export_data_to_excel}
\title{Export data to Excel}
\usage{
iso_export_data_to_excel(
  filepath,
  ...,
  with_explicit_units = TRUE,
  quiet = default(quiet)
)
}
\arguments{
\item{filepath}{the path where to store the Excel file.}

\item{...}{named list of data frames (each will get its own tab)}

\item{with_explicit_units}{whether to include units in all column names}
}
\description{
This is a convenience function to export data to Excel. Use \link{iso_export_calibration_to_excel} for a more specific export function for isoprocessor calibrations.
}
