% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_generate_calibration}
\alias{iso_generate_calibration}
\title{Generate data calibration}
\usage{
iso_generate_calibration(
  dt,
  model,
  calibration = "",
  use_in_calib = default(is_std_peak),
  min_n_datapoints = 2,
  is_std_peak = default(is_std_peak),
  is_standard = default(is_std_peak),
  quiet = default(quiet)
)
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{model}{a single regression model (usually \link[stats]{lm} or \link[stats]{glm}) or a list of multiple alternative regression models for the calibration. If a named list is provided, the name(s) will be used instead of the formulas for the model identification column. If multiple models are provided, the entire data table rows will be duplicated to consider the different models in parallel. Note that \link[stats]{loess} models are supported but discouraged (and will cause a warning) because local polynomial regression fitting does not calibrate based on a hypothesized regression model and can easily mis-calibrate sparse data. The exception is for non-linear temporal drift corrections (use \code{calibration="drift"} to flag as such) which may reasonably require local polynomical regression fitting of the type \code{loess(y ~ file_datetime)} to account for temporal machine variations.}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{use_in_calib}{column or filter condition to determine which subset of data to actually use for the calibration (default is the \code{is_std_peak} field introduced by \code{\link{iso_add_standards}}).}

\item{min_n_datapoints}{the minimum number of data points required for applying the model(s). Note that there is always an additional check to make sure the minimum number of degrees of freedom for each model is met. If the minimum number of degrees of freedom required is not met, the model will/can not be calculated no matter what \code{min_n_datapoints} is set to.}

\item{is_std_peak}{deprecated in favor of \code{use_in_calib}}

\item{is_standard}{deprecated in favor of \code{use_in_calib}}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\value{
the data table with the following columns added (prefixed by the \code{calibration} parameter if provided):
\itemize{
  \item{\code{calib}: }{the name of the calibration if provided in the \code{model} parameter, otherwise the formula}
  \item{\code{calib_ok}: }{a TRUE/FALSE column indicating whether there was enough data for calibration to be generated}
  \item{\code{calib_points}: }{an integer column indicating the total number of data points used in the calibration. Note that this field counts replicate data points: multiple data points that fall at the exact same x or y-value still count as individual data points for this metric.}
  \item{\code{calib_params}: }{a nested dataframe that holds the actual regression model fit, coefficients, summary and data range. These parameters are most easily accessed using the functions \code{\link{iso_unnest_calibration_coefs}}, \code{\link{iso_unnest_calibration_summary}}, \code{\link{iso_unnest_calibration_parameters}}, \code{\link{iso_unnest_calibration_range}}, or directly via \code{\link[tidyr]{unnest}}}
  \item{\code{resid} within \code{all_data}: }{a new column within the nested \code{all_data} that holds the residuals for all standards used in the regression model}
}
}
\description{
Generate a calibration for a specific variable based on one or multiple calibration models. Requires properly nested and grouped data, see \link{iso_prepare_for_calibration} for details. Note that to calibrate different variables, separate calls to this function should be issued each with different \code{calibration} names.
}
\seealso{
Other calibration functions: 
\code{\link{iso_add_standards}()},
\code{\link{iso_apply_calibration}()},
\code{\link{iso_evaluate_calibration_range}()},
\code{\link{iso_get_calibration_coefficients}()},
\code{\link{iso_get_calibration_data}()},
\code{\link{iso_get_calibration_range}()},
\code{\link{iso_get_calibration_summary}()},
\code{\link{iso_get_problematic_calibrations}()},
\code{\link{iso_prepare_for_calibration}()},
\code{\link{iso_remove_problematic_calibrations}()}
}
\concept{calibration functions}
