% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_generate_summary_table}
\alias{iso_generate_summary_table}
\title{Generate an overview table with the data}
\usage{
iso_generate_summary_table(dt, ..., cutoff = 1)
}
\arguments{
\item{dt}{data table, can already have a group_by if so desired}

\item{...}{which data columns to include in data overview}

\item{cutoff}{the minimum number of records per group in order to include the group}
}
\description{
Convenience function to summarize means and standard deviationsfor one or multiple data columns. Use \link[dplyr]{group_by} prior to calling this function to generate the data table for individual subsets of the data. The generated data table always includes an \code{n} column with the number of records per group.
}
