% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_get_calibration_parameters}
\alias{iso_get_calibration_parameters}
\alias{iso_unnest_calibration_parameters}
\title{Get calibration parameters}
\usage{
iso_get_calibration_parameters(dt, calibration = last_calibration(dt),
  select_from_coefs = everything(), select_from_summary = everything(),
  keep_remaining_nested_data = FALSE, keep_other_list_data = FALSE,
  keep_other_calibrations = FALSE)

iso_unnest_calibration_parameters(...)
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{the name of the calibration to get information for, by default the last calibration that was generated}

\item{select_from_coefs}{which columns from the fit coeffiencts to include, supports full dplyr syntax including renaming}

\item{select_from_summary}{which columns from the fit summary to include, supports full dplyr syntax including renaming}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}

\item{keep_other_calibrations}{whether other calibrations (besides the one specified by \code{calibration}) should be kept in the data table}
}
\description{
Retrieve both calibration coefficients (\link{iso_get_calibration_coefs}) and calibration summary (\link{iso_get_calibration_summary}) columns in a single step. Typically this requires specifying \code{select_from_coefs} and \code{select_from_summary} as several columns in the coefficients and summary have the same names.
}
\details{
\code{iso_unnest_calibration_parameters} is deprecated in favor of the equivalent \code{iso_get_calibration_parameters} to standardize function call names. Please start using \code{iso_get_calibration_parameters} as \code{iso_unnest_calibration_parameters} will be deprecated in a future version of isoprocessor.
}
