% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_get_calibration_range}
\alias{iso_get_calibration_range}
\alias{iso_unnest_calibration_range}
\title{Get calibration range}
\usage{
iso_get_calibration_range(dt, calibration = last_calibration(dt),
  select = everything(), keep_remaining_nested_data = FALSE,
  keep_other_list_data = FALSE, keep_other_calibrations = FALSE)

iso_unnest_calibration_range(...)
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{the name of the calibration to get information for, by default the last calibration that was generated}

\item{select}{which columns to unnest - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default (i.e. unnests an entire nested data frame).}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}

\item{keep_other_calibrations}{whether other calibrations (besides the one specified by \code{calibration}) should be kept in the data table}
}
\description{
Retrieve range information created by \link{iso_evaluate_calibration_range}.
}
\details{
\code{iso_unnest_calibration_range} is deprecated in favor of the equivalent \code{iso_get_calibration_range} to standardize function call names. Please start using \code{iso_get_calibration_range} as \code{iso_unnest_calibration_range} will be deprecated in a future version of isoprocessor.
}
