% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_get_default_processor_parameters}
\alias{iso_get_default_processor_parameters}
\title{Get the current default processor parameters}
\usage{
iso_get_default_processor_parameters(data = NULL)
}
\arguments{
\item{data}{can be used to include this function call within a pipeline}
}
\description{
Retrieve a table with all default parameters for the isoprocessor package
(see \link[isoreader]{iso_get_default_reader_parameters} for the equivalent isoreader function).
To set processor parameters, see \code{\link{iso_set_default_processor_parameters}}.
To set messaging and caching parameters see \code{\link[isoreader]{iso_info_messages}}.
For a piping compatible version of this function, see \link{iso_show_default_processor_parameters}.
}
\seealso{
Other settings functions: \code{\link{iso_set_default_process_parameters}},
  \code{\link{iso_show_default_processor_parameters}}
}
