% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_get_problematic_calibrations}
\alias{iso_get_problematic_calibrations}
\title{Fetch problematic calibrations}
\usage{
iso_get_problematic_calibrations(dt, calibration = "",
  select = everything(), quiet = default(quiet))
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{select}{which columns to select for display - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\description{
Fetch data sets that have problematic calibrations (not enough data or another error occured during calibration) for further inspection. This function is typically called after \link{iso_generate_calibration} to inspect problematic data sets. It requires the columns generated by \link{iso_generate_calibration} and the same \code{calibration} parameter used there.
}
