% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_get_problematic_peaks}
\alias{iso_get_problematic_peaks}
\title{Renamed to iso_get_problematic_peak_mappings}
\usage{
iso_get_problematic_peaks(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Renamed to iso_get_problematic_peak_mappings
}
\seealso{
Other peak mapping functions: 
\code{\link{iso_map_peaks}()},
\code{\link{iso_remove_problematic_peak_mappings}()},
\code{\link{iso_summarize_peak_mappings}()}
}
\concept{peak mapping functions}
