% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_get_processor_example}
\alias{iso_get_processor_example}
\alias{iso_get_processor_examples}
\title{Example files}
\usage{
iso_get_processor_example(filename)

iso_get_processor_examples()
}
\arguments{
\item{filename}{the name of the example file for which to retrieve the system path}
}
\description{
The isprocessor package comes with a few example data sets to make it easy to illustrate the functionality.
}
\details{
\code{iso_get_processor_example}: retrieve the path to an isoprocessor example data set

\code{iso_get_processor_examples}: list of all available isoprocessor example data set
}
