% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{iso_identify_outliers}
\alias{iso_identify_outliers}
\title{Identify outliers}
\usage{
iso_identify_outliers(
  dt,
  y,
  plus_minus_value = NULL,
  plus_minus_sd = NULL,
  group_by = NULL,
  is_outlier = default(is_outlier)
)
}
\description{
Identify outliers
}
