% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_mark_calibration_range}
\alias{iso_mark_calibration_range}
\title{Visualize the calibration range}
\usage{
iso_mark_calibration_range(p, calibration = "")
}
\arguments{
\item{p}{a plot generated by \code{\link{iso_plot_data}} (usually piped to this function)}

\item{calibration}{name of the calibration, must match the name used in \link{iso_generate_calibration} (if any)}
}
\description{
This is a convenience function to visualize the calibration ranges in a plot generated by \code{\link{iso_plot_data}} with a gray area behind the data points. Note that \code{\link{iso_evaluate_calibration_range}} must have been called at an earlier point to establish calibration ranges for different terms. Also note that if one of the chosen dimensions (x or y) is not part of the available calibration range for a panel but the other dimension is, it simply shows two lines bracketing the samples in the dimension that is part of the calibration range.
}
