% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_mark_value_range}
\alias{iso_mark_value_range}
\title{Visualize value range}
\usage{
iso_mark_value_range(
  p,
  mean = TRUE,
  plus_minus_value = c(),
  plus_minus_sd = sd,
  sd = 1:2,
  color = NULL
)
}
\arguments{
\item{p}{a \link[ggplot2]{ggplot} object, typically generated by \code{\link{iso_plot_data}}}

\item{mean}{logical (default TRUE), whether to show a line for the value mean}

\item{plus_minus_value}{which fixed value intervals to show (mean +/- fixed), by default none are included.}

\item{plus_minus_sd}{which standard deviation intervals to show (mean +/- sd), by default +/- 1 and +/-2 population std. deviations. To omit these intervals, set \code{sd = c()}.}

\item{sd}{renamed to the more descriptive \code{plus_minus_sd}}

\item{color}{use to override inherited color aesthetics with a fixed value and thus create value ranges that average across all data points in a panel (for this to work, make sure the group aesthetic of the parent plot does NOT include the color aesthetic). Has to be a valid color string, e.g. \code{"black"} or \code{"#ffab05"}. If omitted or \code{color = NULL}, the color aesthetic from the plot is inherited (the default behavior).}
}
\description{
This is a convenience function to visualize value ranges (mean +/- std. deviation) with horizontal lines in a plot, typically one generated by \code{\link{iso_plot_data}}. Considers aesthetics color and group as well as the facet variables (if any are set) to calculate averages within each panel, color and group. This leaves the aesthetics fill and shape to be used for other purposes in displaying data points.
}
\seealso{
Other plot functions: 
\code{\link{iso_mark_calibration_range}()},
\code{\link{iso_mark_outliers}()},
\code{\link{iso_mark_x_range}()},
\code{\link{iso_plot_continuous_flow_data}()},
\code{\link{iso_plot_data}()},
\code{\link{iso_plot_dual_inlet_data}()},
\code{\link{iso_plot_raw_data}()},
\code{\link{iso_plot_ref_peaks}()},
\code{\link{iso_plot_residuals}()},
\code{\link{iso_plot_scan_data}()},
\code{\link{iso_prepare_continuous_flow_plot_data}()},
\code{\link{iso_prepare_dual_inlet_plot_data}()},
\code{\link{iso_prepare_scan_plot_data}()}
}
\concept{plot functions}
