% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_mark_x_range}
\alias{iso_mark_x_range}
\title{Visualize x range}
\usage{
iso_mark_x_range(p, condition, color = NA, fill = "gray", alpha = 0.25)
}
\arguments{
\item{p}{a \link[ggplot2]{ggplot} object, typically generated by \code{\link{iso_plot_data}}}

\item{condition}{any \link[dplyr]{filter} expression to identify x ranges to mark}

\item{color}{color of the shading boxes}

\item{fill}{fill of the shading boxes}

\item{alpha}{alpha of the shading boxes}
}
\description{
This is a convenience function to visualize x ranges with gray vertical boxes in a plot (typically one generated by \code{\link{iso_plot_data}}) by specifying a simple filter condition. Considers the facet variables (if any are set) to evaluate the condition within each panel.
}
\seealso{
Other plot functions: 
\code{\link{iso_mark_calibration_range}()},
\code{\link{iso_mark_outliers}()},
\code{\link{iso_mark_value_range}()},
\code{\link{iso_plot_continuous_flow_data}()},
\code{\link{iso_plot_data}()},
\code{\link{iso_plot_dual_inlet_data}()},
\code{\link{iso_plot_raw_data}()},
\code{\link{iso_plot_ref_peaks}()},
\code{\link{iso_plot_residuals}()},
\code{\link{iso_plot_scan_data}()},
\code{\link{iso_prepare_continuous_flow_plot_data}()},
\code{\link{iso_prepare_dual_inlet_plot_data}()},
\code{\link{iso_prepare_scan_plot_data}()}
}
\concept{plot functions}
