% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_plot_calibration_range}
\alias{iso_plot_calibration_range}
\title{Plot calibration range}
\usage{
iso_plot_calibration_range(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function is deprecated, please use \link{iso_plot_data} and \link{iso_mark_calibration_range} instead.
}
