% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_plot_raw_data}
\alias{iso_plot_raw_data}
\title{Plot raw data from isoreader files}
\usage{
iso_plot_raw_data(iso_files, ..., quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{...}{parameters for the data specific plotting functions}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\description{
Convenience function for making quick standard plots for raw isoreader data.
Calls \code{\link{iso_plot_continuous_flow_data}} and \code{\link{iso_plot_dual_inlet_data}} for data specific plotting (see those functions for parameter details).
For customizing plotting calls, it is easier to use \code{\link{iso_plot_continuous_flow_data}} and \code{\link{iso_plot_dual_inlet_data}} directly.
}
\seealso{
Other plot functions: \code{\link{iso_format}},
  \code{\link{iso_mark_calibration_range}},
  \code{\link{iso_mark_outliers}},
  \code{\link{iso_mark_value_range}},
  \code{\link{iso_plot_continuous_flow_data}},
  \code{\link{iso_plot_data}},
  \code{\link{iso_plot_dual_inlet_data}},
  \code{\link{iso_plot_ref_peaks}},
  \code{\link{iso_plot_residuals}},
  \code{\link{iso_prepare_continuous_flow_plot_data}}
}
\concept{plot functions}
