% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_plot_ref_peaks}
\alias{iso_plot_ref_peaks}
\title{Plot reference peaks}
\usage{
iso_plot_ref_peaks(dt, is_ref_condition, ratio = default(ratio),
  group_id = default(group_id), within_group = FALSE)
}
\arguments{
\item{is_ref_condition}{condition to identify which of the peaks are reference peaks}

\item{ratio}{which ratio column(s) to compare across the reference peaks}

\item{group_id}{group identifier column(s) to clarify which peaks belongs together - first column in group id is used for x axis labels}

\item{within_group}{whether to visualize the deviation within the specified group or across all groups}
}
\description{
Visualize how consistent the reference peaks across a serious of samples.
}
