% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_plot_scan_data}
\alias{iso_plot_scan_data}
\alias{iso_plot_scan_data.iso_file_list}
\alias{iso_plot_scan_data.data.frame}
\title{Plot data from scan files}
\usage{
iso_plot_scan_data(...)

\method{iso_plot_scan_data}{iso_file_list}(
  iso_files,
  data = character(),
  type,
  filter = NULL,
  x_interval = c(),
  y_interval = c(),
  panel = file_id,
  color = data,
  linetype = NULL,
  label = data,
  ...
)

\method{iso_plot_scan_data}{data.frame}(
  df,
  type,
  x_interval = c(),
  y_interval = c(),
  panel = file_id,
  color = data,
  linetype = NULL,
  label = data,
  ...
)
}
\arguments{
\item{...}{additional parameters passed on to \link{iso_plot_data}}

\item{iso_files}{collection of iso_file objects}

\item{data}{which masses and ratios to plot (e.g. \code{c("44", "45", "45/44")} - without the units), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using \code{\link{iso_calculate_ratios}} prior to plotting.}

\item{type}{which type of scan data to plot. Only required if there are more than one type of scan data.}

\item{filter}{any filter condition to apply to the data beyond the masses/ratio selection (param \code{data}). For details on the available data columns see \link[isoreader]{iso_get_raw_data} with parameters \code{gather = TRUE} and \code{include_file_info = everything()} (i.e. all file info is available for plotting aesthetics).}

\item{x_interval}{optional constraints on x axis values}

\item{y_interval}{optional constraints on y axis values}

\item{panel}{whether to panel the data by anything. If using a single parameter (e.g. \code{panel = panel}), will generate a \link[ggplot2]{facet_wrap}. If using a formula (e.g. \code{panel = panel ~ .} or \code{panel = file_id ~ panel}), will generate a \link[ggplot2]{facet_grid}. The default for this parameter is to panel via facet grid by the y variable name but only if multiple \code{y} columns are provided. Otherwise will not generate any facets. If additional facet parameters are desired, please leave use \link[ggplot2]{facet_wrap} and \link[ggplot2]{facet_grid} diretly.}

\item{color}{variable to use for color aesthetic for the plot or constant value for the point and line color}

\item{linetype}{variable to use for linetype aesthetic for the plot or constant value for the line type}

\item{label}{this is primarily of use for turning the generated ggplots into interactive plots via \code{\link[plotly]{ggplotly}} as the \code{label} will be rendered as an additional mousover label.}

\item{df}{a data frame of the scan data prepared for plotting (see \code{\link{iso_prepare_scan_plot_data}})}
}
\description{
This function provides easy plotting for mass and ratio traces from IRMSs scan data. It can be called either directly with a set of \code{iso_file} objects, or with a data frame prepared for plotting scan data (see \code{\link{iso_prepare_scan_plot_data}}).
}
\seealso{
Other plot functions: 
\code{\link{iso_mark_calibration_range}()},
\code{\link{iso_mark_outliers}()},
\code{\link{iso_mark_value_range}()},
\code{\link{iso_mark_x_range}()},
\code{\link{iso_plot_continuous_flow_data}()},
\code{\link{iso_plot_data}()},
\code{\link{iso_plot_dual_inlet_data}()},
\code{\link{iso_plot_raw_data}()},
\code{\link{iso_plot_ref_peaks}()},
\code{\link{iso_plot_residuals}()},
\code{\link{iso_prepare_continuous_flow_plot_data}()},
\code{\link{iso_prepare_dual_inlet_plot_data}()},
\code{\link{iso_prepare_scan_plot_data}()}
}
\concept{plot functions}
