% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_prepare_for_calibration}
\alias{iso_prepare_for_calibration}
\title{Prepare data set for calibration}
\usage{
iso_prepare_for_calibration(dt, group_by = NULL,
  nest_existing_calibrations = FALSE, quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{group_by}{what to group by for indidual calibration calculations (use c(...) to select multiple) - set \code{group_by = NULL} to avoid grouping}

\item{nest_existing_calibrations}{whether to nest existing calibrations. This is usually not necessary and only leads to larger data objects. By default, existing calibrations are therefore never included in the nested data set.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\value{
a nested data set with the \code{group_by} columns out front and the remaining data in a new nested column named \code{all_data}
}
\description{
Nests data set in preparation for calibration calculations. Use the \code{group_by} parameter to group analyses as appropriate for calibration. Use \link{iso_get_calibration_data} to easily pull out columns from the nested data frame in \code{all_data} at any point before, during or after calibration.
}
\seealso{
Other calibration functions: \code{\link{iso_add_standards}},
  \code{\link{iso_apply_calibration}},
  \code{\link{iso_evaluate_calibration_range}},
  \code{\link{iso_generate_calibration}},
  \code{\link{iso_get_calibration_coefficients}},
  \code{\link{iso_get_calibration_data}},
  \code{\link{iso_get_calibration_range}},
  \code{\link{iso_get_calibration_summary}},
  \code{\link{iso_get_problematic_calibrations}},
  \code{\link{iso_remove_problematic_calibrations}}
}
\concept{calibration functions}
