% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_print_data_table}
\alias{iso_print_data_table}
\title{Print data table}
\usage{
iso_print_data_table(
  dt,
  select = everything(),
  filter = TRUE,
  print_func = default(print_func),
  title = NULL,
  unique = TRUE,
  ...
)
}
\arguments{
\item{dt}{data table}

\item{select}{which columns to select (use c(...) to select multiple), supports all \link[dplyr]{select} syntax}

\item{filter}{any filter conditions to apply, by default does not filter any}

\item{print_func}{what function to use for printing (makes it easier to switch between different types of output)}

\item{title}{whether to provide a title message}

\item{unique}{whether to print only unique rows}

\item{...}{additional parameters to forward to the \code{func} function}
}
\value{
the passed in data table (for piping)
}
\description{
This function is DEPRECATED and will be removed in a future version.
}
\note{
this is not working as well as planned - consider removing again or just allowing it to be a simplifying function..
}
