% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_remove_list_columns}
\alias{iso_remove_list_columns}
\title{Remove nested data}
\usage{
iso_remove_list_columns(dt)
}
\description{
Convenience function to remove nested columns in the data table (e.g. in preparation for printing to console or RMarkdown).
}
\note{
not unit tested
}
