% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{iso_remove_missing_metadata}
\alias{iso_remove_missing_metadata}
\title{Remove entries with missing metadata}
\usage{
iso_remove_missing_metadata(dt, remove_has_metadata_column = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{dt}{data with metadata added}

\item{remove_has_metadata_column}{whether to automatically remove the \code{has_metadata} column}
}
\value{
the part of the data table that has metadata, with the \code{has_metadata} removed if \code{remove_has_metadata_column = TRUE}
}
\description{
Continue data processing without data that is missing metadata. This function is typically called after \link{iso_add_metadata} to focus on entries that have metadata. Requires the \code{has_metadata} column to be present.
}
\seealso{
Other metadata functions: \code{\link{iso_add_metadata}},
  \code{\link{iso_get_missing_metadata}}
}
