% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_remove_problematic_calibrations}
\alias{iso_remove_problematic_calibrations}
\title{Remove problematic calibrations}
\usage{
iso_remove_problematic_calibrations(dt,
  calibration = last_calibration(dt), remove_calib_ok_column = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{remove_calib_ok_column}{whether to automatically remove the calibration ok (\code{calib_ok}) column after using it to remove problematic calibrations. This helps automatically clean up the data table and remove information that is no longer needed.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\description{
Remove calibrations that were problematic.
}
\seealso{
Other calibration functions: \code{\link{iso_add_standards}},
  \code{\link{iso_apply_calibration}},
  \code{\link{iso_evaluate_calibration_range}},
  \code{\link{iso_generate_calibration}},
  \code{\link{iso_get_calibration_coefficients}},
  \code{\link{iso_get_calibration_data}},
  \code{\link{iso_get_calibration_range}},
  \code{\link{iso_get_calibration_summary}},
  \code{\link{iso_get_problematic_calibrations}},
  \code{\link{iso_prepare_for_calibration}}
}
\concept{calibration functions}
