% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_remove_problematic_peak_mappings}
\alias{iso_remove_problematic_peak_mappings}
\alias{iso_remove_problematic_peak_mappings.iso_file_list}
\alias{iso_remove_problematic_peak_mappings.data.frame}
\title{Remove problematic peak mappings}
\usage{
iso_remove_problematic_peak_mappings(...)

\method{iso_remove_problematic_peak_mappings}{iso_file_list}(iso_files,
  remove_unidentified = TRUE, remove_missing = TRUE,
  remove_ambiguous = TRUE, remove_mapping_info_columns = TRUE,
  quiet = default(quiet))

\method{iso_remove_problematic_peak_mappings}{data.frame}(peak_table,
  remove_unidentified = TRUE, remove_missing = TRUE,
  remove_ambiguous = TRUE, remove_mapping_info_columns = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{...}{S3 method placeholder parameters, see class specific functions for details on parameters}

\item{iso_files}{collection of continuous flow iso_file objects}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}

\item{peak_table}{data frame with mapped peaks. Requires the \code{is_identified}, \code{is_missing} and \code{is_ambiguous} columns to be present.}

\item{remove_mapping_info_column}{whether to automatically remove mapping info columns. If true and:
\itemize{
 \item{\code{remove_unidentified = TRUE} - }{\code{is_identified} column automatically removed}
 \item{\code{remove_missing = TRUE} - }{\code{is_missing} column automatically removed}
 \item{\code{remove_ambiguous = TRUE} - }{\code{is_ambiguous}, \code{n_matches}, and \code{n_overlapping} columns automatically removed}
}}
}
\value{
the data with the problematic peak mappings remove
}
\description{
Remove peak mappings that were problematic (unidentified, missing, or ambiguous) during peak mapping.
}
\seealso{
Other peak mapping functions: \code{\link{iso_get_problematic_peaks}},
  \code{\link{iso_map_peaks}},
  \code{\link{iso_summarize_peak_mappings}}
}
\concept{peak mapping functions}
