% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{iso_scale_double_with_units}
\alias{iso_scale_double_with_units}
\title{Scale a double with units}
\usage{
iso_scale_double_with_units(x, to_units)
}
\arguments{
\item{x}{an \link[isoreader]{iso_double_with_units} object}

\item{to_units}{what unit to scale to (must be in the same base unit as \code{x})}

\item{return}{\code{x} scaled to the new \code{to_units}}
}
\description{
Scale an isoverse \link[isoreader]{iso_double_with_units} to a different SI prefix. Will error if units are incompatible or si prefix is unknown.
}
