% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_set_default_process_parameters}
\alias{iso_set_default_process_parameters}
\alias{iso_reset_default_process_parameters}
\title{Set default parameters
Define the default parameters for calculations used in this package.}
\usage{
iso_set_default_process_parameters(data = NULL, ...)

iso_reset_default_process_parameters(data = NULL)
}
\arguments{
\item{data}{can be used to include this function call within a pipeline}

\item{...}{named entries for the default parameters (standard or non-standard evaluation supported).}
}
\value{
invisibly returns the passed in \code{data} object
Names are equivalent to function parameter names and
}
\description{
Set default parameters
Define the default parameters for calculations used in this package.
}
\details{
\code{iso_reset_default_process_parameters} resets all default function parameters for this package.
}
\seealso{
Other settings functions: \code{\link{iso_get_default_processor_parameters}},
  \code{\link{iso_show_default_processor_parameters}}
}
