% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table.R
\name{iso_set_peak_table}
\alias{iso_set_peak_table}
\title{Set peak table}
\usage{
iso_set_peak_table(iso_files, peak_table = tibble(file_id =
  character(0)), reset_missing = TRUE, quiet = default(quiet))
}
\arguments{
\item{iso_files}{iso file(s) for which to set peak tables}

\item{peak_table}{peak table data frame, must have a 'file_id' column at minimum (to match with the iso_files)}

\item{reset_missing}{whether to reset the \code{peak_table} of any \code{iso_files} whose\code{file_id} is not in the \code{peak_table}. If set to \code{FALSE}, will preserve pre-existing peak table data instead.}
}
\description{
Set peak table for a collection of isofiles. Peak tables have to have at least a \code{file_id} column but are otherwise flexible. That said, isoprocessor functionality that operates on peak tables assumes that all columns with numbers are \code{\link[isoreader]{iso_double_with_units}} and that retention times are recorded in \code{rt_start}, \code{rt_end}, and \code{rt} (apex retention time) columns.
}
\details{
This funciton will issue a warning if the \code{peak_table} has \code{file_id} values that are not part of the \code{iso_files}. It will reset peak tables for all \code{iso_files} whose \code{file_id} is not part of \code{peak_table} unless \code{reset_missing} is set to \code{FALSE}, in which case pre-existing peak table data will be preserved unaltered. To reset all peak tables, simply call this function without the \code{peak_table} parameter.
}
\seealso{
Other peak table functions: \code{\link{iso_get_peak_table}},
  \code{\link{iso_set_peak_table_from_vendor_data_table}}
}
\concept{peak table functions}
