% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table.R
\name{iso_set_peak_table_from_auto_vendor_data_table}
\alias{iso_set_peak_table_from_auto_vendor_data_table}
\title{renamed to iso_set_peak_table_automatically_from_vendor_data_table}
\usage{
iso_set_peak_table_from_auto_vendor_data_table(...)
}
\description{
This function has been renamed to \link{iso_set_peak_table_automatically_from_vendor_data_table}. Please call \link{iso_set_peak_table_automatically_from_vendor_data_table} directly to avoid the warning.
}
