% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_show_default_processor_parameters}
\alias{iso_show_default_processor_parameters}
\title{Show the current default processor parameters}
\usage{
iso_show_default_processor_parameters(data = NULL,
  print_func = default(print_func), ..., quiet = default(quiet))
}
\arguments{
\item{data}{can be used to include this function call within a pipeline}

\item{...}{additional parameters to forward to the \code{func} function}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}

\item{func}{function to use for formatting the reader parameters table, e.g. \code{\link[knitr]{kable}}.
Note that if the output is in RMarkdown chunks, the chunk option must have \code{results="asis"} for the table to be correctly formatted.}
}
\description{
Shows a table with the default function parameters for this package.
}
\note{
DEPRECATE
}
\seealso{
Other settings functions: \code{\link{iso_get_default_processor_parameters}},
  \code{\link{iso_set_default_processor_parameters}}
}
\concept{settings functions}
