% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_summarize_data_table}
\alias{iso_summarize_data_table}
\title{Summarize a data table}
\usage{
iso_summarize_data_table(dt, ..., cutoff = 1)
}
\arguments{
\item{dt}{data table, can already have a group_by if so desired}

\item{...}{which data columns to include in data overview. All \link[dplyr]{select} style syntax is supported (including on the fly renaming). If no columns are specified, will summarize all numeric columns (excluding any grouping columns).}

\item{cutoff}{the minimum number of records per group in order to include the group}
}
\description{
Convenience function to summarize means and standard deviations for one or multiple data columns. Use \link[dplyr]{group_by} prior to calling this function to generate the data table for individual subsets of the data. The generated data table always includes an \code{n} column with the number of records per group. If no column(s) are specified, will automatically summarize all columns except for the grouping columns.
}
