% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_unnest_calibration_parameters}
\alias{iso_unnest_calibration_parameters}
\title{Unnest calibration parameters}
\usage{
iso_unnest_calibration_parameters(dt, calibration = last_calibration(dt),
  select_from_coefs = everything(), select_from_summary = everything(),
  keep_remaining_nested_data = FALSE, keep_other_list_data = TRUE)
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{select_from_coefs}{which columns from the fit coeffiencts to include, supports full dplyr syntax including renaming}

\item{select_from_summary}{which columns from the fit summary to include, supports full dplyr syntax including renaming}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}
}
\description{
Convenience function to unnest both calibration coefficients (\link{iso_unnest_calibration_coefs}) and calibration summary (\link{iso_unnest_calibration_summary}) columns in a single step.
}
