% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_unnest_calibration_summary}
\alias{iso_unnest_calibration_summary}
\title{Unnest calibration summary}
\usage{
iso_unnest_calibration_summary(dt, calibration = "", select = everything(),
  keep_remaining_nested_data = FALSE, keep_other_list_data = TRUE)
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{select}{which columns to unnest - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default (i.e. unnests an entire nested data frame).}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}
}
\description{
Retrieve summary entries for a calibration. Problematic calibrations are silently omitted (use \link{iso_get_problematic_calibtations} and \link{iso_remove_problematic_calibrations} to deal with them more explicitly).
}
