% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_unnest_data}
\alias{iso_unnest_data}
\title{Unnest data}
\usage{
iso_unnest_data(dt, select = everything(),
  keep_remaining_nested_data = TRUE, keep_other_list_data = TRUE)
}
\arguments{
\item{select}{which columns to unnest - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default (i.e. unnests an entire nested data frame).}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}
}
\description{
Pull out columns from the \code{all_data} column. Typically used at various points after a dataset is prepared for calibration with \link{iso_prepare_for_calibration}, calibrations are generated with \link{iso_generate_calibration} and/or calibrations are applied with \link{iso_apply_calibration}. Note that unnesting of data columns that have multiple values per data set leads to row replication.
}
