% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_unnest_delta_calib_coefs}
\alias{iso_unnest_delta_calib_coefs}
\alias{iso_unnest_delta_calib_summary}
\alias{iso_unnest_calib_data}
\title{Retrieve calibration information}
\usage{
iso_unnest_delta_calib_coefs(dt, select = everything(),
  keep_remaining_nested_data = FALSE, keep_other_list_data = TRUE)

iso_unnest_delta_calib_summary(dt, select = everything(),
  keep_remaining_nested_data = FALSE, keep_other_list_data = TRUE)

iso_unnest_calib_data(dt, select = everything(),
  calibration_data = default(calibration_data),
  keep_remaining_nested_data = TRUE, keep_other_list_data = TRUE)
}
\arguments{
\item{select}{which fields from the selected model result column to unnest}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}
}
\description{
Retrieve calibration information
}
\details{
\code{iso_unnest_delta_calib_coefs} unnests the regression fit coefficients for the delta calibration

\code{iso_unnest_delta_calib_summary} unnests the regression fit summary for the delta calibration

\code{iso_unnest_calib_data} unnests the data
}
