% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_visualize_data}
\alias{iso_visualize_data}
\title{Visualize the data}
\usage{
iso_visualize_data(dt, x, y, group = NULL, color = NULL, shape = NULL,
  size = NULL, linetype = NULL, lines = TRUE, points = FALSE)
}
\arguments{
\item{dt}{data frame with peak data}

\item{x}{the x axis for the calibration parameters, by default a datetime but can also work as text}

\item{y}{which columns to visualize}

\item{group}{what to group by, multiple columns allowed (combine with c(...)), usually not necessary if color or other grouping is defined}

\item{lines}{whether to plot lines}

\item{points}{whether to plot points}
}
\description{
General purpose visualization function.
}
\note{
should probably make sure that the default columns for gather 'panel' and 'value' do not exist...

it would be great to allow renaming of the columns via this (especially the y column)
}
