% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_visualize_delta_calib_fits}
\alias{iso_visualize_delta_calib_fits}
\title{Visualize results for delta calibration}
\usage{
iso_visualize_delta_calib_fits(dt, x = default(file_datetime), color = NULL,
  shape = NULL, size = NULL, date_breaks = "2 hours",
  date_labels = "\%d \%b \%H:\%M", include_from_summary = c(adj.r.squared,
  deviance))
}
\arguments{
\item{dt}{data table}

\item{x}{the x axis for the calibration parameters, by default a datetime but can also work as text}

\item{date_breaks}{what breaks to use for the x axis if it is a datetime}

\item{date_labels}{datetime label pattern for x axis if it is a datetime}

\item{include_from_summary}{which parameters from the fit summary to include}
}
\description{
Visualize results for delta calibration
}
