% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_visualize_ref_peaks}
\alias{iso_visualize_ref_peaks}
\title{Visualize reference peaks}
\usage{
iso_visualize_ref_peaks(dt, is_ref_condition, ratio = default(ratio),
  group_id = default(group_id), within_group = FALSE)
}
\arguments{
\item{group_id}{first column in group id is used for x axis}

\item{within_group}{whether to visualize the deviation within the specified group or across all groups}
}
\description{
Visualize reference peaks
}
\note{
: refine and write tests
}
