% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoprocessor-package}
\alias{isoprocessor}
\alias{isoprocessor-package}
\title{isoprocessor: IRMS data processor}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Data processing and reduction pipelines for isotope ratio mass spectrometry (IRMS) data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/isoprocessor}
  \item Report bugs at \url{https://github.com/kopflab/isoprocessor/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
