% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoprocessorCUB-package}
\alias{isoprocessorCUB}
\alias{isoprocessorCUB-package}
\title{isoprocessorCUB: CU Boulder IRMS data processor}
\description{
CU Boulder specific version of the standard IRMS data processor
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/isoprocessorCUB}
  \item Report bugs at \url{https://github.com/kopflab/isoprocessorCUB/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
