% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nest_data}
\alias{nest_data}
\title{nest data based on the grouping
this is basically just an inverse version (in terms of selection) of nest that can handle the sophisticated selection parameters of select}
\usage{
nest_data(dt, group_by = NULL, nested_data = default(nested_data))
}
\arguments{
\item{group_by}{what to keep out of the nested data (i.e. what to group by), by default nests everything}
}
\description{
nest data based on the grouping
this is basically just an inverse version (in terms of selection) of nest that can handle the sophisticated selection parameters of select
}
