% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H3_factor.R
\name{plot_H3_factors}
\alias{plot_H3_factors}
\title{Plot H3 factor}
\usage{
plot_H3_factors(
  dt,
  file_datetime = default(file_datetime),
  H3_factor = default(H3_factor),
  is_H3_factor_file = default(is_H3_factor_file),
  color = low_amp,
  label = default(file_id),
  date_breaks = "12 hours",
  date_labels = "\%b \%d - \%H:\%m"
)
}
\arguments{
\item{dt}{the data table}
}
\description{
Plot H3 factor
}
