% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_data_table}
\alias{print_data_table}
\title{Print data table}
\usage{
print_data_table(dt, select = everything(), filter = TRUE, func = NULL,
  ...)
}
\arguments{
\item{dt}{data table}

\item{select}{which columns to select (use c(...) to select multiple), supports all \link[dplyr]{select} syntax}

\item{filter}{any filter conditions to apply, by default does not filter any}

\item{func}{function to use for formatting the reader parameters table, e.g. \code{\link[knitr]{kable}}.
Note that if the output is in RMarkdown chunks, the chunk option must have \code{results="asis"} for the table to be correctly formatted.}

\item{...}{additional parameters to forward to the \code{func} function}
}
\value{
the passed in data table (for piping)
}
\description{
Convenience function to print a data table as part of a pipe.
}
