% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H3_factor.R
\name{print_H3_factors}
\alias{print_H3_factors}
\title{Print H3 factor summary}
\usage{
print_H3_factors(dt, file_id = default(file_id),
  file_datetime = default(file_datetime), H3_factor = default(H3_factor),
  ...)
}
\description{
Print H3 factor summary
}
