% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{run_regression}
\alias{run_regression}
\title{run a set of regressions}
\usage{
run_regression(dt, model, model_data = default(model_data),
  model_filter_condition = default(), model_name = default(model_name),
  model_fit = default(model_fit), model_coefs = default(model_coefs),
  model_summary = default(model_summary), residual = default(residual),
  min_n_data_points = 2)
}
\arguments{
\item{dt}{data table}

\item{model}{the regression model or named list of regression models.
If a named list is provided, the name(s) will be stored in the \code{model_name} column}

\item{model_data}{the nested model data column}

\item{model_filter_condition}{a filter to apply to the data before running the regression, by default no filter}

\item{model_name}{new column with the model names if any are supplied}

\item{model_fit}{the new model objects column}

\item{model_coefs}{the new model coefficients nested data frame column}

\item{model_summary}{the new model summary nested data frame column}

\item{residual}{the new residual column in the nested model_data}

\item{min_n_data_points}{- the minimum number of data points required for the regression}

\item{residuals}{resid name of the residuals column}
}
\description{
run a set of regressions
}
