% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_default_parameters}
\alias{set_default_parameters}
\alias{reset_default_parameters}
\alias{get_default_parameters}
\alias{show_default_parameters}
\title{Set default parameters
Define the default parameters for calculations used in this package.}
\usage{
set_default_parameters(data = NULL, ...)

reset_default_parameters(data = NULL)

get_default_parameters()

show_default_parameters(data = NULL)
}
\arguments{
\item{data}{can be used to include this function call within a pipeline}

\item{...}{named entries for the default parameters (standard or non-standard evaluation supported).}
}
\value{
invisibly returns the passed in \code{data} object
Names are equivalent to function parameter names and
}
\description{
Set default parameters
Define the default parameters for calculations used in this package.
}
\details{
\code{reset_default_parameters} resets all default function parameters for this package.

\code{get_default_parameters} returns all default function parameters for this package.

\code{show_default_parameters} shows a table with the default function parameters for this package.
}
