% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H3_factor.R
\name{summarize_H3_factors}
\alias{summarize_H3_factors}
\title{Summarize H3 factors}
\usage{
summarize_H3_factors(
  dt,
  is_H3_factor_file_condition,
  is_H3_factor_file = default(is_H3_factor_file),
  file_id = default(file_id),
  file_datetime = default(file_datetime),
  H3_factor = default(H3_factor),
  ampl = default(ampl),
  keep = default(NULL)
)
}
\arguments{
\item{dt}{data table}

\item{is_H3_factor_file_condition}{condition to find H3 factor files}
}
\description{
Summarize H3 factors
}
