% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unnest_model_column}
\alias{unnest_model_column}
\title{Unnest model results column}
\usage{
unnest_model_column(
  dt,
  model_column,
  model_params = model_params,
  nested_model = FALSE,
  select = everything(),
  keep_remaining_nested_data = FALSE,
  keep_other_list_data = FALSE
)
}
\arguments{
\item{model_column}{name of the model column to unnest}

\item{model_params}{name of the model params column that holds all the other model columns (if \code{nested_model = TRUE})}

\item{nested_model}{whether the model is nested, if TRUE, must also provide \code{model_params}}

\item{select}{which columns to unnest - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default (i.e. unnests an entire nested data frame).}

\item{keep_remaining_nested_data}{whether to keep any remaining parts of the partially unnested data (irrelevant if \code{select = everything()})}

\item{keep_other_list_data}{keep other list data columns (e.g. other data or model columns)}
}
\description{
Convenience functions for unnesting model columns (supports both if regression is stored nested or unnested - see \link{run_regression} for details).
}
