% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_regression}
\alias{apply_regression}
\title{invert a regression for calibration purposes}
\usage{
apply_regression(dt, predict, nested_model = FALSE,
  calculate_error = FALSE, model_data = model_data,
  model_name = model_name, model_fit = model_fit,
  model_params = model_params, predict_value = pred,
  predict_error = pred_se, predict_range = NULL)
}
\arguments{
\item{dt}{data table with calibrations}

\item{predict}{which value to calculate, must be one of the regression's independent variables}

\item{nested_model}{whether the model is nested, if TRUE, must also provide \code{model_params}}

\item{calculate_error}{whether to estimate the standard error from the calibration (using the Wald method), stores the result in the new \code{predict_error} column. Note that error calculation slows this function down a fair bit and is therefore disabled by default.}

\item{model_data}{the nested model data column}

\item{model_name}{new column with the model formulae or names if supplied}

\item{model_fit}{the new model objects column}

\item{model_params}{the nested model information (only relevant if \code{nest_model = TRUE})}

\item{predict_value}{the new column in the model_data that holds the predicted value}

\item{predict_error}{the new column in the model_data that holds the error of the predicted value (only created if \code{calculate_error = TRUE})}

\item{predict_range}{vector of 2 numbers, if provided will be used for finding the solution for the predict variable. By default uses the range observed in the calibration variables. Specifying the \code{predict_range} is usually only necessary if the calibration range should be extrapolated significantely.}
}
\description{
note: this is optimized for inverting regressions, but should also do prediction for regular regression at some point
}
