% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_add_standards}
\alias{iso_add_standards}
\title{Add calibration standards}
\usage{
iso_add_standards(dt, stds, match_by = default(compound),
  is_standard = default(is_standard), quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{stds}{the standards table}

\item{match_by}{default is the compound parameter also used in peak mapping}

\item{is_standard}{new column that holds information about what is a standard and what isn't}
}
\description{
Add calibration standards
}
\note{
FIXME write unit tests
}
