% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_add_standards}
\alias{iso_add_standards}
\alias{iso_add_standards.iso_file_list}
\alias{iso_add_standards.data.frame}
\title{Add calibration standards}
\usage{
iso_add_standards(...)

\method{iso_add_standards}{iso_file_list}(
  iso_files,
  stds,
  match_by = default(std_match_by),
  is_std_peak = default(is_std_peak),
  quiet = default(quiet)
)

\method{iso_add_standards}{data.frame}(
  dt,
  stds,
  match_by = default(std_match_by),
  is_std_peak = default(is_std_peak),
  quiet = default(quiet),
  is_standard = NULL
)
}
\arguments{
\item{...}{S3 method placeholder parameters, see class specific functions for details on parameters}

\item{iso_files}{collection of continuous flow iso_file objects}

\item{stds}{standards data frame}

\item{match_by}{what column(s) to match the standards by}

\item{is_std_peak}{new column that holds information about which ones are standard peaks (i.e. have known isotopic values)}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}

\item{dt}{data frame with the peak table}

\item{is_standard}{renamed to \code{is_std_peak} because the naming caused too much confusion, will be removed in future versions, please use \code{is_std_peak} instead}
}
\value{
iso files with standards data frame merged into the peak table and the following information column added:
\itemize{
\item{\code{is_std_peak}: }{a logical TRUE/FALSE indicating which data table entry is a standard}
}
}
\description{
Convenience function to add calibration standards to a continuous flow peak table (standalone or inside an isofiles object).
}
\seealso{
Other calibration functions: 
\code{\link{iso_apply_calibration}()},
\code{\link{iso_evaluate_calibration_range}()},
\code{\link{iso_generate_calibration}()},
\code{\link{iso_get_calibration_coefficients}()},
\code{\link{iso_get_calibration_data}()},
\code{\link{iso_get_calibration_range}()},
\code{\link{iso_get_calibration_summary}()},
\code{\link{iso_get_problematic_calibrations}()},
\code{\link{iso_prepare_for_calibration}()},
\code{\link{iso_remove_problematic_calibrations}()}
}
\concept{calibration functions}
