% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_apply_calibration}
\alias{iso_apply_calibration}
\title{Apply calibration}
\usage{
iso_apply_calibration(dt, predict, calibration = "",
  predict_range = NULL, calculate_error = FALSE,
  quiet = default(quiet))
}
\arguments{
\item{dt}{nested data table with \code{all_data} and calibration columns (see \link{iso_generate_calibration})}

\item{predict}{which value to calculate, must be one of the regression's independent variables}

\item{calibration}{name of the calibration to apply, must match the name used in \link{iso_generate_calibration} (if any)}

\item{predict_range}{vector of 2 numbers, if provided will be used for finding the solution for the predict variable. By default uses the range observed in the calibration variables. Specifying the \code{predict_range} is usually only necessary if the calibration range should be extrapolated significantely.}

\item{calculate_error}{whether to estimate the standard error from the calibration (using the Wald method), stores the result in the new \code{predict_error} column. Note that error calculation slows this function down a fair bit and is therefore disabled by default.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\value{
the data table with the following columns added to the nested \code{all_data} \:
\itemize{
  \item{\code{predict} column with suffix \code{_pred}: }{the predicted value from applying the calibration}
  \item{\code{predict} column with suffix \code{_pred_se}: }{the error of the predicated value propagated from the calibration. Only created if \code{calculate_error = TRUE}.}
  \item{\code{predict} column with suffix \code{_pred_in_range}: }{reports whether a data entry is within the range of the calibration by checking whether ALL dependent and independent variables in the regression model are within the range of the calibration - is set to FALSE if any(!) of them are not - i.e. this column provides information on whether new values are extrapolated beyond a calibration model and treat the extrapolated ones with the appropriate care. Note that all missing predicted values (due to missing parameters) are also automatically flagged as not in range}
}
}
\description{
Apply calibration
}
