% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_calibration_to_excel}
\alias{iso_export_calibration_to_excel}
\title{Export calibration to Excel}
\usage{
iso_export_calibration_to_excel(
  calibs_df,
  filepath,
  ...,
  include_all_data = TRUE,
  include_calib_coefs = TRUE,
  include_calib_summary = TRUE,
  include_calib_range = TRUE,
  with_explicit_units = TRUE,
  quiet = default(quiet)
)
}
\arguments{
\item{calibs_df}{data frame that includes calibration columns (i.e. the output of functions like \link{iso_generate_calibration}, \link{iso_apply_calibration}, \link{iso_evaluate_calibration_range}). Can be the output from \link{iso_get_calibration_data} if called after a calibration has been generated.}

\item{filepath}{the path where to store the Excel file.}

\item{...}{named list of additional data frames to include in the summary (each will get their own tab)}

\item{include_all_data}{whether to include a tab for all the peak data - if this function is called after \link{iso_apply_calibration} this will include}

\item{include_calib_coefs}{whether to include a tab for the model coefficients}

\item{include_calib_summary}{whether to include a tab for the model summary}

\item{include_calib_range}{whether to include a tab for the evaluation range (if available, i.e. generated by \link{iso_evaluate_calibration_range})}

\item{with_explicit_units}{whether to include units in all column names}
}
\value{
returns the \code{calib} object invisibly for use in pipelines
}
\description{
This function exports the passed in calibration(s) data frame to Excel. The different kinds of information (all data, model coefficients, summary, evaluation range) are exported to separate tabs within the excel file. If there are multiple different calibrations in the data frame (i.e. multiple \code{*calib_params} columns), ALL will be exported with different tabs for each.
}
