% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_generate_calibration}
\alias{iso_generate_calibration}
\title{Generate data calibration}
\usage{
iso_generate_calibration(dt, model, calibration = "",
  is_standard = default(is_standard), min_n_datapoints = 2,
  quiet = default(quiet))
}
\arguments{
\item{dt}{nested data table with column \code{all_data} (see \link{iso_prepare_for_calibration})}

\item{model}{a single regression model or a list of multiple alternative regression models for the calibration. If a named list is provided, the name(s) will be used instead of the formulaes for the model identification column. Note that if multiple models are provided, the entire data table rows will be duplicated to consider the different models in parallel.}

\item{calibration}{an informative name for the calibration (could be e.g. \code{"d13C"} or \code{"conc"}). If provided, will be used as a prefix for the new columns generated by this function. This parameter is most useful if there are multiple variables in the data set that need to be calibrated (e.g. multiple delta values, concentration, etc.). If there is only a single variable to calibrate, the \code{calibration} parameter is completely optional and can just be left blank (the default).}

\item{is_standard}{column or filter condition to determine which subset of data to actually use for the calibration (default is the \code{is_standard} field introduced by \code{\link{iso_add_standards}}).}
}
\value{
the data table with the following columns added (prefixed by the \code{calibration} parameter if provided):
\itemize{
  \item{\code{calib}: }{the name of the calibration if provided in the \code{model} parameter, otherwise the formula}
  \item{\code{calib_ok}: }{a TRUE/FALSE column indicating whether there was enough data for calibration to be generated}
  \item{\code{calib_params}: }{a nested dataframe that holds the actual regression model fit, coefficients, summary and data range. These parameters are most easily accessed using the functions \code{\link{iso_unnest_calibration_coefs}}, \code{\link{iso_unnest_calibration_summary}}, \code{\link{iso_unnest_calibration_parameters}}, \code{\link{iso_unnest_calibration_range}}, or directly via \code{\link[tidyr]{unnest}}}
  \item{\code{resid} within \code{all_data}: }{a new column within the nested \code{all_data} that holds the residuals for all standards used in the regression model}
}
}
\description{
Generate a calibration for a specific variable based on one or multiple calibration models. Requires properly nested and grouped data, see \link{iso_prepare_for_calibration} for details. Note that to calibrate different variables, separate calls to this function should be issued each with different \code{calibration} names.
}
