% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{iso_generate_summary_table}
\alias{iso_generate_summary_table}
\title{Alias for summarize data table}
\usage{
iso_generate_summary_table(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Alias for summarize data table
}
