% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_get_problematic_peak_mappings}
\alias{iso_get_problematic_peak_mappings}
\title{Fetch problematic peak mappings}
\usage{
iso_get_problematic_peak_mappings(dt, select = everything(),
  unidentified = TRUE, missing = TRUE, ambiguous = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{dt}{data table with mapped peaks. Requires the \code{is_identified}, \code{is_missing} and \code{is_ambiguous} columns to be present.}

\item{select}{which columns to select - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. Includes all columns by default but this function is more useful with a smaller subset of identifying columns.}

\item{unidentified}{whether to include peaks that are problematics because they are unidentified}

\item{missing}{whether to include peaks that are problematics because they are missing}

\item{ambiguous}{whether to include peaks that are problematics because they are ambiguously identified}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\value{
data table with rows for problematic peaks and the \code{select}-identified columns
}
\description{
Fetch peak mappings that were problematic in any way. This function is typically called after \link{iso_map_peaks} to inspect problematic entries. Use the \code{select} parameter to select only the most informative columns (always includes at minimum the \code{peak_info} and \code{problem} columns to identify why the peak is problematic). Note that peaks that are ambiguous because of multiple potential map matches have a data table entry for each potential match.
}
\seealso{
Other peak mapping functions: \code{\link{iso_map_peaks}},
  \code{\link{iso_remove_problematic_peak_mappings}},
  \code{\link{iso_summarize_peak_mappings}}
}
\concept{peak mapping functions}
