% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_get_problematic_peaks}
\alias{iso_get_problematic_peaks}
\title{Renamed to iso_get_problematic_peak_mappings}
\usage{
iso_get_problematic_peaks(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Renamed to iso_get_problematic_peak_mappings
}
