% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_map_peaks}
\alias{iso_map_peaks}
\alias{iso_map_peaks.iso_file_list}
\alias{iso_map_peaks.data.frame}
\title{Map peaks based on retention time}
\usage{
iso_map_peaks(...)

\method{iso_map_peaks}{iso_file_list}(iso_files, peak_maps,
  map_id = default(map_id), compound = default(compound),
  rt = default(rt), rt_start = default(rt_start),
  rt_end = default(rt_end), rt_prefix_divider = ":",
  quiet = default(quiet))

\method{iso_map_peaks}{data.frame}(peak_table, peak_maps,
  file_id = default(file_id), map_id = default(map_id),
  compound = default(compound), rt = default(rt),
  rt_start = default(rt_start), rt_end = default(rt_end),
  rt_prefix_divider = ":", quiet = default(quiet))
}
\arguments{
\item{...}{S3 method placeholder parameters, see class specific functions for details on parameters}

\item{iso_files}{collection of continuous flow iso_file objects}

\item{peak_maps}{data frame with the peak map(s). At minimum, this data frame must have a \code{compound} and \code{rt} column but may have additional information columns. If multiple peak maps are provided, the \code{peak_table} data frame requires a \code{map_id} column to identify which peak map should be used and the peak maps data frame must have a \code{rt:<map_id>} column for each used value of \code{map_id}. The names of all these columns can be changed if necessary using the \code{compound}, code{rt} and \code{map_id} parameters.}

\item{map_id}{the column in peak_table that indicates which map to use for which file (only necessary if multiple peak maps are used)}

\item{compound}{the column in peak_maps that holds compound information}

\item{rt}{the column in peak_table and colum prefix in peak_maps ("rt:...") that holds retention time information}

\item{rt_start}{the column in peak_table that holds start of peak retention times}

\item{rt_end}{the column in peak_table that holds end of peak retention times}

\item{rt_prefix_divider}{the divider after the retention time column prefix in peak_maps to identify the map id values (e.g. "rt:map_id_value")}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}

\item{peak_table}{data frame with the peak table}

\item{file_id}{the column(s) in peak_table that uniquely identify a file/set of peaks that belong together}
}
\value{
data frame with mapped peaks and the following information columns:
\itemize{
\item{\code{peak_info}: }{a label for the peak with its name and retention time plus indicators of any ambiguity in identification in the form of \code{'?'} for either compound name or retention time for an expected peak that was not found}
\item{\code{is_identified}: }{a logical TRUE/FALSE indicating peaks that have been successfully identified (includes missing peaks from the peak map!) (note that this information could also be derived from !is.na(compound) but is provided for convenience)}
\item{\code{is_missing}: }{a logical TRUE/FALSE indicating peaks that are in the peak map definition but have no matching peak}
\item{\code{is_ambiguous}: }{a logical TRUE/FALSE indicating peaks that are ambiguous in their definition either because they have multiple matches or because they overlap with other, overlapping peaks that were identified the same (note that this information could also be derived from n_overlapping > 1 | n_matches > 1 but is provided for convenience)}
\item{\code{n_matches}: }{the number of matches each peak has in the peak map}
\item{\code{n_overlapping}: }{the number of overlapping peaks that match the same peak definition}
}
}
\description{
This function makes it easy to map peaks based on pre-defined peak maps. It is implemented to work on \code{iso_files} as well as directly on \code{peak_table} data frames. It reports all peaks including missing peaks and ambiguous peaks by adding a set of information columns for each entry (\code{is_identified}, \code{is_missing}, \code{is_ambiguous}, \code{n_matches}, \code{n_overlapping}). For routine downstream data processing, this function is usually followed by \code{\link{iso_summarize_peak_mappings}} and \code{\link{iso_get_problematic_peak_mappings}} to inspect the problematic peaks and \code{\link{iso_remove_problematic_peak_mappings}} to proceed only with mapped peaks that are clearly identified. Note that without this filter, one must proceed with great caution interpreting the ambiguous peaks. Also note that if the \code{compound} column already exists in \code{peak_table}, it will be overwritten with the new mappings from the peak maps but will issue a warning that this is happening.
}
\seealso{
Other peak mapping functions: \code{\link{iso_get_problematic_peaks}},
  \code{\link{iso_remove_problematic_peak_mappings}},
  \code{\link{iso_summarize_peak_mappings}}
}
\concept{peak mapping functions}
