% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_mark_outliers}
\alias{iso_mark_outliers}
\title{Mark outliers}
\usage{
iso_mark_outliers(
  p,
  condition = NULL,
  plus_minus_value = NULL,
  plus_minus_sd = sd_cutoff,
  sd_cutoff = NULL,
  label = NULL,
  size = 2,
  shape = 16,
  color = "black"
)
}
\arguments{
\item{p}{a \link[ggplot2]{ggplot} object, typically generated by \code{\link{iso_plot_data}}}

\item{condition}{any \link[dplyr]{filter} expression to identify outliers}

\item{plus_minus_value}{value cutoff to identify outliers (mean +/- value)}

\item{plus_minus_sd}{standard deviation cutoff to identify outliers (mean +/- sd)}

\item{sd_cutoff}{renamed to the more descriptive \code{plus_minus_sd}}

\item{label}{optional expression to add a label to the outlier points}

\item{size}{size of the outlier points}

\item{shape}{shape of the outlier points}

\item{color}{color of the outlier points}
}
\description{
This is a convenience function to visually highlight values that fall outside the data range via a flexible \code{condition} statement, \code{plus_minus_value} cutoff from the mean, or \code{plus_minus_sd} standard deviation cutoff from the mean. Optionally with a text label to make it easier to identify the analysis.
}
\seealso{
Other plot functions: 
\code{\link{iso_mark_calibration_range}()},
\code{\link{iso_mark_value_range}()},
\code{\link{iso_plot_continuous_flow_data}()},
\code{\link{iso_plot_data}()},
\code{\link{iso_plot_dual_inlet_data}()},
\code{\link{iso_plot_raw_data}()},
\code{\link{iso_plot_ref_peaks}()},
\code{\link{iso_plot_residuals}()},
\code{\link{iso_plot_scan_data}()},
\code{\link{iso_prepare_continuous_flow_plot_data}()},
\code{\link{iso_prepare_scan_plot_data}()}
}
\concept{plot functions}
