% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_mark_value_range}
\alias{iso_mark_value_range}
\title{Visualize value range}
\usage{
iso_mark_value_range(p, mean = TRUE, sd = 1:2)
}
\arguments{
\item{p}{a \link[ggplot2]{ggplot} object, typically generated by \code{\link{iso_plot_data}}}

\item{mean}{logical (default TRUE), whether to show a linef or the value mean}

\item{sd}{which standard deviation intervals to show (mean +/- sd), by default +/- 1 and +/-2 population std. deviations. To omit these intervals, set \code{sd = c()}.}
}
\description{
This is a convenience function to visualize value ranges (mean +/- std. deviation) with horizontal lines in a plot, typically one generated by \code{\link{iso_plot_data}}. Considers aesthetics color and group as well as the facet variables to calculate averages within each panel, color and group. This leaves aesthetics fill and shape to be used for other purposes in displaying data points.
}
\seealso{
Other plot functions: 
\code{\link{iso_format}()},
\code{\link{iso_mark_calibration_range}()},
\code{\link{iso_mark_outliers}()},
\code{\link{iso_plot_continuous_flow_data}()},
\code{\link{iso_plot_data}()},
\code{\link{iso_plot_dual_inlet_data}()},
\code{\link{iso_plot_raw_data}()},
\code{\link{iso_plot_ref_peaks}()},
\code{\link{iso_plot_residuals}()},
\code{\link{iso_prepare_continuous_flow_plot_data}()}
}
\concept{plot functions}
