% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table.R
\name{iso_mutate_peak_table}
\alias{iso_mutate_peak_table}
\title{Mutate peak table}
\usage{
iso_mutate_peak_table(iso_files, ..., quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{...}{dplyr-style \link[dplyr]{mutate} conditions applied to the combined peak_table (see \code{\link{iso_get_peak_table}})}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\description{
Mutate the peak table (\code{\link{iso_get_peak_table}}) within isofile objects by changing existing columns or introducing new ones. Works just like dplyr's \link[dplyr]{mutate}.
}
\seealso{
Other peak table functions: \code{\link{iso_get_peak_table}},
  \code{\link{iso_set_peak_table_from_vendor_data_table}},
  \code{\link{iso_set_peak_table}}
}
\concept{peak table functions}
