% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_plot_data}
\alias{iso_plot_data}
\title{Visualize the data}
\usage{
iso_plot_data(dt, x, y, y_error = NULL, group = NULL, color = NULL,
  fill = NULL, shape = NULL, size = 4, linetype = NULL,
  label = NULL, lines = FALSE, points = FALSE)
}
\arguments{
\item{dt}{data frame with peak data}

\item{x}{the x axis aesthetic for the calibration parameters, can be datetime, text or numeric}

\item{y}{which columns to visualize, combine with c()}

\item{y_error}{an error column for drawing y error bars - if multiple \code{y} are provided, error needs to point to the same number of columns}

\item{group}{what to group by, multiple columns allowed (combine with paste(...)), usually not necessary if groupings are fully defined through other aesthetics}

\item{color}{variable to use for color aesthetic for the plot}

\item{fill}{variable to use for the fill aesthetic of the plot}

\item{shape}{variable to use for shape aesthetic for the plot}

\item{size}{variable to use for size aesthetic for the plot or constant value for the points size}

\item{linetype}{variable to use for linetype aesthetic for the plot}

\item{label}{this is primarily of use for turning the generated ggplots into interactive plots via \code{\link[plotly]{ggplotly}} as the \code{label} will be rendered as an additional mousover label.}

\item{lines}{whether to plot lines (FALSE by default)}

\item{points}{whether to plot points (FALSE by default)}
}
\description{
General purpose convenience visualization function. Simply add other ggplot components after calling this function to customize more (e.g. with \link[ggplot2]{facet_wrap} or \link[ggplot2]{theme} calls). Make sure to specify \code{lines = TRUE} and/or \code{points = TRUE} to add the layers.
}
\note{
should probably make sure that the default columns for gather 'panel' and 'value' do not exist...

it would be great to allow renaming of the columns via this (especially the y column)
}
