% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{iso_prepare_continuous_flow_plot_data}
\alias{iso_prepare_continuous_flow_plot_data}
\title{Prepare plotting data from continuous flow files}
\usage{
iso_prepare_continuous_flow_plot_data(iso_files, data = c(),
  include_file_info = NULL, time_interval = c(),
  time_interval_units = "seconds", filter = NULL, normalize = FALSE,
  zoom = NULL, zoom_group = data)
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{data}{which masses and ratios to plot (e.g. \code{c("44", "45", "45/44")} - without the units), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using \code{\link{iso_calculate_ratios}} prior to plotting.}

\item{time_interval}{which time interval to plot}

\item{time_interval_units}{which units the time interval is in, default is "seconds"}

\item{filter}{any filter condition to apply to the data beyond the masses/ratio selection (param \code{data}) and time interval (param \code{time_interval}). For details on the available data columns see \link{iso_get_raw_data} with parameters \code{gather = TRUE} and \code{include_file_info = everything()} (i.e. all file info is available for plotting aesthetics).}

\item{normalize}{whether to normalize the data (default is FALSE, i.e. no normalization). If TRUE, normalizes each trace across all files (i.e. normalized to the global max/min). This is particularly useful for overlay plotting different mass and/or ratio traces (\code{panel = NULL}). Note that zooming (if \code{zoom} is set) is applied after normalizing.}

\item{zoom}{if not set, automatically scales to the maximum range in the selected time_interval in each plotting panel. If set, scales by the indicated factor, i.e. values > 1 are zoom in, values < 1 are zoom out, baseline always remains the bottom anchor point. Note that zooming is always relative to the max in each zoom_group (by default \code{zoom_group = data}, i.e. each trace is zoomed separately). The maximum considered may be outside the visible time window. Note that for \code{zoom_group} other than \code{data} (e.g. \code{file_id} or \code{NULL}), zooming is relative to the max signal across all mass traces. Typically it makes most sense to set the \code{zoom_group} to the same variable as the planned \code{panel} parameter to the plotting function. Lastly, note that zooming only affects masses, ratios are never zoomed.}
}
\description{
This functions helps with the preparation of plotting data from continuous flow data files (i.e. the raw chromatogram data). Call either explicity and pass the result to \code{\link{iso_plot_continuous_flow_data}} or let \code{\link{iso_plot_continuous_flow_data}} take care of preparing the plotting data directly from the \code{iso_files}.
}
\seealso{
Other plot functions: \code{\link{iso_plot_continuous_flow_data}},
  \code{\link{iso_plot_dual_inlet_data}},
  \code{\link{iso_plot_raw_data}}
}
\concept{plot functions}
