% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_prepare_for_calibration}
\alias{iso_prepare_for_calibration}
\title{Prepare data set for calibration}
\usage{
iso_prepare_for_calibration(dt, group_by = NULL,
  quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{group_by}{what to group by for indidual calibration calculations (use c(...) to select multiple) - set \code{group_by = NULL} to avoid grouping}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages.}
}
\value{
a nested data set with the \code{group_by} columns out front and the remaining data in a new nested column named \code{all_data}
}
\description{
Nests data set in preparation for calibration calculations. Use the \code{group_by} parameter to group analyses as appropriate for calibration. Use \link{iso_unnest_data} to easily pull out additional columns from the nested data frame in \code{all_data} at any point before, during or after calibration.
}
