% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{iso_prepare_for_calibration}
\alias{iso_prepare_for_calibration}
\title{Prepare data set for calibration}
\usage{
iso_prepare_for_calibration(dt, group_by = default(group_by),
  calibration_data = default(calibration_data), quiet = default(quiet))
}
\arguments{
\item{dt}{data table}

\item{group_by}{what to group by for indidual calibration calculations (use c(...) to select multiple) - set \code{group_by = NULL} to avoid grouping}

\item{calibration_data}{column name for the combined calibration data}

\item{stds}{the standards table}

\item{keep}{which column(s) to keep beyond the group_by (use c(...) to select multiple). Note that these should be unique across grouping variables, otherwise it will result in multiple lines of the same regression data per group.}
}
\description{
Prepare data set for calibration
}
\note{
FIXME write unit tests
}
