% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_mapping.R
\name{iso_summarize_peak_mappings}
\alias{iso_summarize_peak_mappings}
\title{Summarize peaks}
\usage{
iso_summarize_peak_mappings(dt, file_id = default(file_id))
}
\arguments{
\item{dt}{data frame with peak data}

\item{file_id}{the column(s) in dt that uniquely identify a file/set of peaks that belong together}
}
\value{
summary data table with one row for each unique combination of the \code{file_id} parameter, the number of expected peaks (\code{expected}), found peaks (\code{found}), and problematic peaks (\code{problematic}), as well as a long text column (\code{peak_info}) listing all the peaks and their retention times
}
\description{
Summarize peaks after peak mapping. This function is called after \link{iso_map_peaks} and can be used in combination with \link{iso_get_problematic_peaks} to inspect problematic peaks in particular.
}
\seealso{
Other peak mapping functions: \code{\link{iso_get_problematic_peaks}},
  \code{\link{iso_map_peaks}},
  \code{\link{iso_remove_problematic_peaks}}
}
\concept{peak mapping functions}
