% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R, R/settings.R, R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{extract_substring}
\alias{reexports}
\alias{extract_word}
\alias{reexports}
\alias{parse_number}
\alias{reexports}
\alias{parse_logical}
\alias{reexports}
\alias{parse_integer}
\alias{reexports}
\alias{parse_double}
\alias{reexports}
\alias{parse_datetime}
\alias{reexports}
\alias{iso_turn_info_messages_on}
\alias{reexports}
\alias{iso_turn_info_messages_off}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{unnest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{isoreader}{\code{\link[isoreader]{extract_substring}}, \code{\link[isoreader]{extract_word}}, \code{\link[isoreader]{iso_turn_info_messages_on}}, \code{\link[isoreader]{iso_turn_info_messages_off}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{parse_number}}, \code{\link[readr]{parse_logical}}, \code{\link[readr]{parse_integer}}, \code{\link[readr]{parse_double}}, \code{\link[readr]{parse_datetime}}}

  \item{tidyr}{\code{\link[tidyr]{unnest}}}
}}

